feat: 홍보 카테고리랑 글 합치기
