+++ src/components/AD/index.js
@@ -2,4 +2,3 @@ export { default as ADArticleModal } from "./ADArticleModal";
 export { default as ADCard } from "./ADCard";
 export { default as ADCardWrapper } from "./ADCardWrapper";
 export { default as ADDateCarousel } from "./ADDateCarousel";
-export { default as ADCategory } from "./ADCategory";
